Added detection when a localmessagebroadcastpartner's mailslot has become invalid.
In that case, the partner that sees this happen will send a goodbye message on behalf of the dead partner (that has probably crashed or hasn't been exited properly). It will also clean up its own records of course.

Since Wyphon is built upon LocalMessageBroadcast, this will work for Wyphon too.